requirements.txt
# FastAPI and Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery[redis]==5.3.6

# Validation and Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# AWS S3 / File Upload
boto3==1.34.34
python-magic==0.4.27
Pillow==10.2.0

# OpenAI
openai==1.12.0

# Firebase Push Notifications
firebase-admin==6.4.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging and Monitoring
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.40.0

# Rate Limiting
slowapi==0.1.9

